+++ mge_v18_student_version/src/mge/materials/LitTextureMaterial.cpp
@@ -82,7 +82,7 @@ void LitTextureMaterial::render(World* pWorld, Mesh* pMesh, const glm::mat4& pMo
 	std::vector<Light*> _lights = LitMaterial::getLights();
 
 	//pass in the light properties
-	for (int i = 0; i < _lights.size(); i++)
+	for (size_t i = 0; i < _lights.size(); i++)
 	{
 		if (!_overrideSpecularLight)
 		{
